First round of converting entities to Impl entities
