+++ src/main/java/com/mewna/catnip/shard/DispatchEmitter.java
@@ -1,16 +1,8 @@
 package com.mewna.catnip.shard;
 
-import com.google.common.collect.ImmutableList;
 import com.mewna.catnip.Catnip;
 import com.mewna.catnip.entity.EntityBuilder;
-import com.mewna.catnip.entity.Member;
-import com.mewna.catnip.entity.Message;
-import com.mewna.catnip.entity.Message.MessageType;
-import com.mewna.catnip.entity.User;
 import io.vertx.core.json.JsonObject;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 import static com.mewna.catnip.shard.DiscordEvent.*;
